agents modeling/Agents/Image Detection Agent/requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
Pillow==10.2.0
langgraph==0.0.19
